pyproject.toml
[project]
name = "smart-workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=4.2.20",
    "django-browser-reload>=1.21.0",
    "django-tailwind[cookiecutter,honcho,reload]>=4.4.2",
    "djlint>=1.36.4",
    "google-generativeai==0.7.2",
    "jinja2>=3.1.6",
    "pillow==10.2.0",
    "pypdf2==3.0.1",
    "python-docx==1.1.0",
    "python-dotenv==1.0.0",
    "requests>=2.32.5",
]
